TCLD - [fix] refactor Work module
interchange sql raw by stored procedure
