+++ QUANGHANH2/EntityResult/GetDataWork_Result.cs
@@ -1,11 +1,12 @@
-﻿using System;
+﻿using QUANGHANH2.Models;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Web;
 
-namespace QUANGHANH2.Models.HumanResources
+namespace QUANGHANH2.EntityResult
 {
-    public class Work_Extend : Work
+    public class GetDataWork_Result : Work
     {
         public string pay_table { get; set; }
     }
