+++ QUANGHANH2/QUANGHANH2.csproj
@@ -524,7 +524,7 @@
     <Compile Include="Models\HistoryWork.cs" />
     <Compile Include="Models\HoSoNhanVien.cs" />
     <Compile Include="Models\HumanResources\WorkGroup_Extend.cs" />
-    <Compile Include="Models\HumanResources\Work_Extend.cs" />
+    <Compile Include="EntityResult\GetDataWork_Result.cs" />
     <Compile Include="Models\Important_Documentary.cs">
       <DependentUpon>QuangHanhManufacturingEntities.tt</DependentUpon>
     </Compile>
